Fix Whitaker Dylint findings ahead of suite adoption
Prepare the codebase for the Whitaker Dylint suite by resolving every
finding it reports:

- Return `Result` from test fixtures and setup helpers instead of
  calling `expect`, propagating arrangement failures to the test body,
  which now unwraps them as its verdict (`no_expect_outside_tests`).
- Convert shared assertion helpers (`assert_paths`, `assert_cached_match`,
  `assert_try_use_cache`) into macros so panic line numbers continue to
  point at the calling test.
- Split oversized modules below the 400-line ceiling
  (`module_max_lines`): move the async lifecycle API into
  `cluster/async_api.rs`, the async invoker into
  `worker_invoker/async_invoker.rs`, root-privileged preparation into
  `bootstrap/prepare/root.rs`, and the `fs` unit tests into
  `fs_tests.rs`.
- Add inner doc comments to inline test modules
  (`module_must_have_inner_docs`).
- Flatten nested conditional clusters in `shutdown_timeout_from_env`,
  `handle_outcome`, and `env_for_profile` by extracting helpers
  (`bumpy_road_function`).
- Replace panicking `unwrap_or_else` closures in the test-support lock
  helpers with explicit matches (`no_unwrap_or_else_panic`).
